backend/requirements.txt
fastapi==0.104.1
pydantic==2.4.2
pydantic-settings==2.0.3
uvicorn==0.23.2
sqlalchemy==2.0.22
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
pytest==7.4.3
httpx==0.25.1
requests==2.31.0
authlib==1.2.1
opencv-python==4.8.1.78
pytesseract==0.3.10
pillow==10.1.0
alembic==1.12.1
email-validator==2.1.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==19.13.0
